Create Prims.c
I attempted all the algorithms in C because I am more familiar with C language. Also ,I know data structures only in this language. 
Prim's algorithm doesn't work well for all the test cases since it doesn't consider the directed graphs. It considers the graph to be undirected and finally
gives the total cost of the traversal of MST(minimum spanning tree).